archive/old_code/core/session_manager.py
import os
import json
import sqlite3
import asyncio
from cryptography.fernet import Fernet
from typing import Dict, Any

class SessionManager:
    def __init__(self, db_path: str, encryption_key: str):
        self.db_path = db_path
        self.encryption_key = encryption_key
        self.cipher = Fernet(encryption_key)
        self.sessions = {}
        self._load_sessions()

    def _load_sessions(self):
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        cursor.execute("SELECT platform, encrypted_cookies, metadata FROM sessions")
        rows = cursor.fetchall()
        for platform, encrypted_cookies, metadata in rows:
            cookies = self.cipher.decrypt(encrypted_cookies.encode()).decode()
            self.sessions[platform] = {
                "cookies": cookies,
                "metadata": json.loads(metadata),
                "is_valid": True
            }
        conn.close()

    def _save_sessions(self):
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        cursor.execute("DELETE FROM sessions")
        for platform, session in self.sessions.items():
            encrypted_cookies = self.cipher.encrypt(session["cookies"].encode()).decode()
            metadata = json.dumps(session["metadata"])
            cursor.execute("INSERT INTO sessions (platform, encrypted_cookies, metadata) VALUES (?, ?, ?)",
                           (platform, encrypted_cookies, metadata))
        conn.commit()
        conn.close()

    def add_session(self, platform: str, cookies: str, metadata: Dict[str, Any]):
        if platform not in self.sessions:
            self.sessions[platform] = {"cookies": cookies, "metadata": metadata, "is_valid": True}
        else:
            self.sessions[platform]["cookies"] = cookies
            self.sessions[platform]["metadata"] = metadata
        self._save_sessions()

    async def validate_session(self, session: Dict[str, Any]) -> bool:
        # Implement async validation logic here
        # For example, make a request to the platform's API and check the response
        # Return True if the session is still valid, False otherwise
        pass

    def get_session(self, platform: str) -> Dict[str, Any]:
        session = self.sessions.get(platform)
        if session and session["is_valid"]:
            return session
        elif platform in self.sessions:
            self.rotate_session(platform)
            return self.sessions[platform]
        else:
            raise ValueError(f"No session found for platform: {platform}")

    def rotate_session(self, platform: str):
        # Implement session rotation logic here
        # For example, switch to the next available session or create a new one
        pass

    def export_sessions(self) -> str:
        return json.dumps({platform: session["metadata"] for platform, session in self.sessions.items()})

    def import_sessions(self, data: str):
        sessions_data = json.loads(data)
        for platform, metadata in sessions_data.items():
            self.add_session(platform, "", metadata)

    def mark_unhealthy(self, platform: str):
        if platform in self.sessions:
            self.sessions[platform]["is_valid"] = False

# Usage example:
# encryption_key = os.environ.get("ENCRYPTION_KEY")
# session_manager = SessionManager("sessions.db", encryption_key)
# session_manager.add_session("LinkedIn", "cookies", {"user_id": 123})
# session = session_manager.get_session("LinkedIn")