requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-multipart==0.0.6

# Async HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Database
aiosqlite==0.19.0

# Authentication
PyJWT==2.8.0

# AI Service
openai==1.10.0

# Browser Automation - Stagehand AI-Powered Framework by BrowserBase
stagehand-py>=2.0.0
browserbase>=0.3.0
playwright>=1.42.1
httpx>=0.24.0
python-dotenv>=1.0.0

# File Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Typing
typing-extensions>=4.0.0

# Job Scraping
python-jobspy>=1.1.0
